% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{addCellMarkersToCGS}
\alias{addCellMarkersToCGS}
\title{Add gene symbols and cell type marker information to cluster gene statistics}
\usage{
addCellMarkersToCGS(sCV, cellMarkersU, cellMarkersS, symbolMap)
}
\arguments{
\item{sCV}{An object of class \code{sCVdata}.}

\item{cellMarkersU}{Derived from the \code{cellMarkers} argument to
\code{\link{runShiny}}. A list of the unique gene symbols for each cell
type in \code{cellMarkers}.}

\item{cellMarkersS}{Derived from the \code{cellMarkers} argument to
\code{\link{runShiny}}. A list of the gene symbols common to two or more
cell types in \code{cellMarkers}. Each entry is named for the indicies of
\code{cellMarkers} that share the gene.}

\item{symbolMap}{The output of \code{\link{map2symbol}}.}
}
\value{
The \code{sCVdata} object with the new variables added to
  \code{ClustGeneStats}.
}
\description{
Adds four new variables to the cluster-wise gene statistics dataframe of the
scClustViz data object. Official gene symbols are added as variable
\code{genes}. The remaining variables are used in
\code{\link{plot_clusterGenes_markers}} to plot cell type marker genes.
Variables \code{cMu} and \code{cMs} are logical vectors indicating genes that
are unique to and shared across cell type markers respectively. Variable
\code{overCut} indicates which genes should be labelled in the plot.
}
\seealso{
\code{\link{findKeyType}}, \code{\link{map2symbol}}, and
  \code{\link[AnnotationDbi]{mapIds}}.
}
