% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{CalcAllSCV}
\alias{CalcAllSCV}
\title{Prepare all cluster solutions for visualization with scClustViz}
\usage{
CalcAllSCV(inD, clusterDF, assayType = "", DRforClust = "pca",
  exponent = 2, pseudocount = 1, DRthresh = 0.1, testAll = TRUE,
  FDRthresh = 0.05, storeAllDE = T, calcSil = T, calcDEvsRest = T,
  calcDEcombn = T, UseBiocParallel = F)
}
\arguments{
\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}

\item{clusterDF}{A data frame of cluster assignments for all cells in the
dataset. Variables (columns) are cluster solutions with different
parameters, and rows should correspond to cells of the input gene
expression matrix.}

\item{assayType}{Default = NULL (for Seurat v1/2). A length-one character
vector representing the assay slot in which the expression data is stored
in the input object. This is not required for Seurat v1 or v2 objects. See
\code{\link{getExpr}} for details.}

\item{DRforClust}{Default = "pca".A length-one character vector representing
the dimensionality reduction method used as the input for clustering. This
is commonly PCA, and should correspond to the slot name of the cell
embedding in your input data - either the \code{type} argument in
\code{\link[SingleCellExperiment]{reducedDim}(x,type)} or the
\code{reduction.type} argument in
\code{\link[Seurat]{GetDimReduction}(object,reduction.type)} (v2) or
\code{reduction} in \code{\link[Seurat]{Embeddings}(object,reduction)}.}

\item{exponent}{Default = 2. A length-one numeric vector representing the
base of the log-normalized gene expression data to be processed. Generally
gene expression data is transformed into log2 space when normalizing (set
this to 2), though \code{Seurat} uses the natural log (set this to exp(1)).}

\item{pseudocount}{Default = 1. A length-one numeric vector representing the
pseudocount added to all log-normalized values in your input data. Most
methods use a pseudocount of 1 to eliminate log(0) errors.}

\item{DRthresh}{Default = 0.1. A length-one numeric vector between 0 and 1
representing the detection rate threshold for inclusion of a gene in the
differential expression testing. A gene will be included if it is detected
in at least this proportion of cells in at least one of the clusters being
compared.}

\item{testAll}{Default = TRUE. Logical value indicating whether to test all
cluster solutions (\code{TRUE}) or stop testing once a cluster solution has
been found where there is no differentially expressed genes found between
at least one pair of nearest neighbouring clusters (\code{FALSE}). If set
to FALSE, this function will test cluster solutions in ascending order of
number of clusters found. \emph{If set to (\code{FALSE}), only tested
cluster solutions will appear in the scClustViz shiny app.}}

\item{FDRthresh}{Default = 0.05. A length-one numeric vector representing the
targeted false discovery rate used to determine the number of
differentially expressed genes between nearest neighbouring clusters,
assuming \code{testAll} is set FALSE If \code{testAll} is TRUE, this
argument is unused.}

\item{storeAllDE}{Default = TRUE. A logical vector of length 1 indicating
whether to calculate and store effect size information for all genes in the
comparison (TRUE), or just those passing the detection rate threshold for
the Wilcoxon rank-sum test (FALSE). Setting this to FALSE will reduce the
size of the output sCVdata object.}

\item{calcSil}{Default = TRUE. A logical vector of length 1. If TRUE,
silhouette widths (a cluster cohesion/separation metric) will be calculated
for all cells. This calculation is performed using the function
\code{\link{CalcSilhouette}}, which is a wrapper to
\code{\link[cluster]{silhouette}} with distance calculated using the same
reduced dimensional cell embedding as was used for clustering, as indicated
in the \code{DRforClust} argument. If the package \code{cluster} is not
installed, this calculation is skipped.}

\item{calcDEvsRest}{Default = TRUE. A logical vector of length 1. If TRUE,
differential expression tests will be performed comparing each cluster to
the remaining cells in the data using a Wilcoxon rank-sum test and
reporting false discovery rates. This calculation is performed using the
function \code{\link{CalcDEvsRest}}. If set to FALSE, it is suggested that
you perform DE testing on the same set of comparisons using a statistical
method of your choice. This can be passed into your \code{sCVdata} objects
in the list returned by \code{CalcAllSCV} using the function
\code{\link{CalcDEvsRest}}. See function documentation for details.}

\item{calcDEcombn}{Default = TRUE.  A logical vector of length 1. If TRUE,
differential expression tests will be performed comparing all pairwise
combinations of clusters using a Wilcoxon rank-sum test and reporting false
discovery rates. This calculation is performed using the function
\code{\link{calcDEcombn}}. If set to FALSE, it is suggested that you
perform DE testing on the same set of comparisons using a statistical
method of your choice. This can be passed into your \code{sCVdata} objects
in the list returned by \code{CalcAllSCV} using the function
\code{\link{calcDEcombn}}. See function documentation for details.}

\item{UseBiocParallel}{Default = FALSE. Very experimental implementation of
BiocParallel for calculations. Not recommended.}
}
\value{
The function returns a list containing \code{\link{sCVdata}} objects
  for each cluster resolution (sample) in the \code{clusterDF} data frame.
  The output object and the \code{inD} object should be saved as an
  \code{.RData} file. That file is the input for \code{\link{runShiny}}, the
  scClustViz Shiny interaction visualization app. See example. For details of
  calculations performed / stored by this function, see
  \code{\link{sCVdata}}.
}
\description{
An all-in-one function to prepare your data for viewing in the interactive
Shiny app. See example for the basic usage of scClustViz.
}
\details{
This is a wrapper function for running \code{\link{CalcSCV}} over each
cluster resolution in the input, and outputs a list of \code{\link{sCVdata}}
objects that should be saved along with the input data. The resulting file is
ready to be read by \code{\link{runShiny}} for viewing. For each cluster
solution provided, this function calculates summary statistics per gene per
cluster, differential gene expression, and cluster separation metrics. This
may take a while to run, depending on the number of cluster solutions tested.
Use the \code{testAll} argument to prevent testing of overfitted cluster
solutions. To help track its progress, this function uses progress bars from
\code{pbapply}. To disable these, set
\code{\link[pbapply]{pboptions}(type="none")}. To re-enable, set
\code{\link[pbapply]{pboptions}(type="timer")}.
}
\examples{
\dontrun{
your_cluster_columns <- grepl("res[.0-9]+$",
                              names(getMD(your_scRNAseq_data_object)))
# ^ Finds the cluster columns of the metadata in a Seurat object.

your_cluster_results <- getMD(your_scRNAseq_data_object)[your_cluster_columns]

sCVdata_list <- CalcAllSCV(inD=your_scRNAseq_data_object,
                           clusterDF=your_cluster_results,
                           assayData=NULL,
                           DRforClust="pca",
                           exponent=exp(1),
                           pseudocount=1,
                           DRthresh=0.1,
                           testAll=F,
                           FDRthresh=0.05,
                           calcSil=T,
                           calcDEvsRest=T,
                           calcDEcombn=T)

save(your_scRNAseq_data_object,sCVdata_list,
     file="for_scClustViz.RData")

runShiny(filePath="for_scClustViz.RData")
# ^ see ?runShiny for detailed argument list
}

}
\seealso{
\code{\link{sCVdata}} for information on the output data class.
  \code{\link{CalcSCV}} to generate an \code{sCVdata} object for a single
  cluster solution. \code{\link{runShiny}} starts the interactive Shiny GUI
  to view the results of this testing.
}
