% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{CalcDEcombn}
\alias{CalcDEcombn}
\alias{CalcDEcombn,sCVdata-method}
\title{Performs DE testing between pairs of clusters in sCVdata}
\usage{
CalcDEcombn(sCVd, inD, storeAllDE, UseBiocParallel)

\S4method{CalcDEcombn}{sCVdata}(sCVd, inD, storeAllDE = FALSE,
  UseBiocParallel = FALSE)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}

\item{storeAllDE}{A logical vector of length 1 indicating whether to
calculate and store effect size information for all genes in the comparison
(TRUE), or just those passing the detection rate threshold for the Wilcoxon
rank-sum test (FALSE). Setting this to FALSE will reduce the size of the
output sCVdata object.}

\item{UseBiocParallel}{Default = FALSE. Very experimental implementation of
BiocParallel for calculations. Not recommended.}
}
\value{
A named list of data frames, one entry for each pairwise combination
  of levels in \code{Clusters(sCVd)} (with corresponding name where levels
  are separated by '-'). Each entry is data frame containing gene
  differential expression stats when comparing the cells of that cluster to
  all other cells in the input data. Rows represent genes, and variables
  include \code{logGER} (an effect size measure: gene expression ratio in log
  space, often referred to as logFC), \code{dDR} (an effect size measure:
  difference in detection rate), and \code{FDR} (significance measure: false
  discovery rate). Also included are \code{Wstat} and \code{pVal}, the test
  statistic and the p-value of the Wilcoxon rank sum test.
}
\description{
Performs differential gene expression tests between each pairwise combination
of cluster in an sCVdata object using the gene expression matrix of input
data object. Alternatively, this function can be skipped, and existing DE
test results can be assigned directly to the sCVdata object.
}
\details{
This function performs Wilcoxon rank sum tests comparing gene expression
between the cells of all pairwise combinations of cluster clusters in the
input data. Gene expression ratio in log space (\code{logGER}) and
differences in detection rates (\code{dDR}) are reported for all genes in the
comparison. Genes are tested if they are detected in at least one of the
cluster at a higher proportion than \code{Param(sCVd,"DRthresh")}, and both
unadjusted p-values and false discovery rates are reported for all genes
tested. To help track its progress, this function uses progress bars from
\code{pbapply}. To disable these, set
\code{\link[pbapply]{pboptions}(type="none")}. To re-enable, set
\code{\link[pbapply]{pboptions}(type="timer")}.

If using existing DE test results, assign results of differential gene
expression tests for all pairwise combinations of clusters in sCVdata to the
\code{\link{DEcombn}} slot of the \code{\link{sCVdata}} object. See example
and slot documentation.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Calculate DE between cluster pairs
}}

\examples{
\dontrun{
## Example using CalcDEvsRest ##
DEcombn(your_sCV_obj) <- CalcDEcombn(sCVd=your_sCV_obj,
                                     inD=your_scRNAseq_data_object)


## Example using MAST results from Seurat to replace CalcDEcombn ##

MAST_pw <- apply(combn(levels(your_seurat_obj@ident),2),
                 MARGIN=2,
                 function(X) {
                   FindMarkers(your_seurat_obj,
                               ident.1=X[1],
                               ident.2=X[2],
                               logfc.threshold=0,
                               min.pct=0.1,
                               test.use="MAST",
                               latent.vars="nUMI")
                 })
names(MAST_pw) <- apply(combn(levels(your_seurat_obj@ident),2),2,
                        function(X) paste(X,collapse="-"))
# ^ Names must be in "X-Y" format

for (i in names(MAST_pw)) {
  MAST_pw[[i]]$dDR <- MAST_pw[[i]]$pct.1 - MAST_pw[[i]]$pct.2
  # ^ Diff in detect rate (dDR) must be a variable in each dataframe
  names(MAST_pw[[i]])[names(MAST_pw[[i]]) == "avg_logFC"] <- "logGER"
  # ^ Effect size variable must be named 'logGER'
  names(MAST_pw[[i]])[names(MAST_pw[[i]]) == "p_val_adj"] <- "FDR"
  # ^ Significance variable must be named 'FDR'
  # Note: rownames of each dataframe must be gene names, 
  # but FindMarkers should already do this.
}
DEcombn(your_sCV_obj) <- MAST_pw
# ^ Slot MAST results into sCVdata object

}

}
\seealso{
\code{\link{CalcSCV}} for wrapper function to calculate all
  statistics for an sCVdata object, and \code{\link{fx_calcEScombn}} and
  \code{\link{fx_calcDEcombn}} for the internal functions performing the
  calculations. Wilcox test is now powered by \code{\link[presto]{wilcoxauc}}
  for super speed.
}
