% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcDEvsCombn}
\alias{calcDEvsCombn}
\title{Do deVS calculation}
\usage{
calcDEvsCombn(nge, cl, deMes, threshType, FDRthresh)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{deMes}{The output from \code{\link{deMarkerGenesUsed}}.}

\item{threshType}{Filtering genes for use in differential expression testing
can be done multiple ways. We use an expression ratio filter for comparing
each cluster to the rest of the tissue as a whole, but find that difference
in detection rates works better when comparing clusters to each other. You
can set threshType to \code{"logGER"} to use a gene expression ratio for all
gene filtering, or leave it as default (\code{"dDR"}) to use difference in
detection rate as the thresholding method when comparing clusters to each
other.}

\item{FDRthresh}{The false discovery rate to use as a threshold for
determining statistical significance of differential expression calculated
by the Wilcoxon rank-sum test.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
 gene in each cluster to that gene in every other cluster in a series of
 tests. The results are stored as a nested list of dataframes. Each list
 element contains a named list of clusters (cluster A). Each of those lists
 contains a named list of all the other clusters (cluster B). Each of those
 list elements contains a dataframe of four variables, where each sample is a
 gene. \code{dDR} is the difference in detection rate of that gene between
 the two clusters (DR[A] - DR[B]). \code{logGER} is the log gene expression
 ratio calculated by taking the difference in mean expression of the gene
 (see \link{meanLogX} for mean calculation) between the two clusters (MTC[A]
 - MTC[B]). \code{pVal} is the p-value of the Wilcoxon rank sum test.
 \code{qVal} is the false discovery rate-corrected p-value of the test.
}
\description{
Calculates Wilcoxon rank-sum tests for all genes in each of the potential
combinations of clusters to compare. You probably don't need to use this
unless you're trying to customize \code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{deMarkerGenesUsed}} for the preceding calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
