% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcDEvsRest}
\alias{calcDEvsRest}
\title{Do deTissue calculation}
\usage{
calcDEvsRest(nge, cl, deT_effectSize, exponent, pseudocount, FDRthresh)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{deT_effectSize}{The output from \code{\link{calcESvsRest}}.}

\item{exponent}{The log base of your normalized input data. Seurat
normalization uses the natural log (set this to exp(1)), while other
normalization methods generally use log2 (set this to 2).}

\item{pseudocount}{The pseudocount added to all log-normalized values in your
input data. Most methods use a pseudocount of 1 to eliminate log(0) errors.}

\item{FDRthresh}{The false discovery rate to use as a threshold for
determining statistical significance of differential expression calculated
by the Wilcoxon rank-sum test.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
 gene in each cluster to the rest of the cells as a whole in a one vs all
 comparison. The results are stored as a named list of dataframes. There is a
 list element for each cluster containing a dataframe of three variables,
 where each sample is a gene. \code{logGER} is the log gene expression ratio
 calculated by subtracting the mean expression of the gene (see
 \link{meanLogX} for mean calculation) in all other cells from the mean
 expression of the gene in this cluster. \code{pVal} is the p-value of the
 Wilcoxon rank sum test. \code{qVal} is the false discovery rate-corrected
 p-value of the test.
}
\description{
Calculates Wilcoxon rank-sum tests for all genes in each one-vs-all comparison
of a cluster vs the rest of the data. You probably don't need to use this
unless you're trying to customize \code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{calcESvsRest}} for the preceding calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
