% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcDist}
\alias{calcDist}
\title{Do deDist calculation}
\usage{
calcDist(deVS)
}
\arguments{
\item{deVS}{The output from \code{\link{calcDEvsCombn}}.}
}
\value{
A matrix of distances between clusters for each cluster resolution.
 Distances are calculated as number of differentially expressed genes between
 clusters. Interpretable by \code{\link{dist}} to generate a \code{dist}
 object.
}
\description{
Counts number of DE genes between each pair of clusters from
\code{\link{calcDEvsCombn}} output. You probably don't need to use this unless
you're trying to customize \code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{calcDEvsCombn}} for the preceding calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
