% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcESvsCombn}
\alias{calcESvsCombn}
\title{Calculate genes used for deVS calculation}
\usage{
calcESvsCombn(cl, CGS, threshType, logGERthresh, dDRthresh)
}
\arguments{
\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{CGS}{The output from \code{\link{calcCGS}}.}

\item{threshType}{Filtering genes for use in differential expression testing
can be done multiple ways. We use an expression ratio filter for comparing
each cluster to the rest of the tissue as a whole, but find that difference
in detection rates works better when comparing clusters to each other. You
can set threshType to \code{"logGER"} to use a gene expression ratio for all
gene filtering, or leave it as default (\code{"dDR"}) to use difference in
detection rate as the thresholding method when comparing clusters to each
other.}

\item{logGERthresh}{Magnitude of gene expression ratio for a gene between
clusters to use as filter for determining which genes to test for
differential expression between clusters.}

\item{dDRthresh}{Magnitude of detection rate difference of a gene between
clusters to use as filter for determining which genes to test for
differential expression between clusters.}
}
\value{
The function returns a list where each list element is a dataframe
 with effect size statistics (log-ratios of gene expression and difference in
 detection rate) when comparing each gene in a cluster to the rest of the
 cells as a whole in a one vs all comparison. These dataframes are filtered
 to only include those gene that pass the relevant threshold, and thus the
 rownames for each list entry correspond to the genes to test in
 \code{\link{calcDEvsCombn}}.
}
\description{
Calculates the log-ratios of gene expression or difference in detection rate
for all genes in each of the potential combinations of clusters to compare.
This is used to determine the genes used in deVS calculations. You probably
don't need to use this unless you're trying to customize
\code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{calcDEvsCombn}} for the followup DE calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
