% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcMarker}
\alias{calcMarker}
\title{Do deMarker calculation}
\usage{
calcMarker(deVS, FDRthresh)
}
\arguments{
\item{deVS}{The output from \code{\link{calcDEvsCombn}}.}

\item{FDRthresh}{The false discovery rate to use as a threshold for
determining statistical significance of differential expression calculated
by the Wilcoxon rank-sum test.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
 gene in each cluster to that gene in every other cluster in a series of
 tests, and filtering for only those genes that show significant positive
 differential expression versus all other clusters. The results are stored as
 a list of dataframes. Each list element contains a named list of clusters
 (cluster A). Each of those list elements contains a dataframe where
 variables represent comparisons to all the other clusters and each sample is
 a gene. For each other cluster (cluster B), there are three variables, named
 as follows: \code{vs.B.dDR} is the difference in detection rate of that gene
 between the two clusters (DR[A] - DR[B]). \code{vs.B.logGER} is the log gene
 expression ratio calculated by taking the difference in mean expression of
 the gene (see \link{meanLogX} for mean calculation) between the two clusters
 (MTC[A] - MTC[B]). \code{vs.B.qVal} is the false discovery rate-corrected
 p-value of the Wilcoxon rank sum test.
}
\description{
Identifies marker genes for each cluster from \code{\link{calcDEvsCombn}} output.
You probably don't need to use this unless you're trying to customize
\code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{calcDEvsCombn}} for the preceding calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
