% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{calcNeighb}
\alias{calcNeighb}
\title{Do deNieghb calculation}
\usage{
calcNeighb(deVS, deDist)
}
\arguments{
\item{deVS}{The output from \code{\link{calcDEvsCombn}}.}

\item{deDist}{The output from \code{\link{deDist}}.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
 gene in each cluster to that gene in its nearest neighbouring cluster
 (calculated by number of differentially expressed genes), and filtering for
 only those genes that show significant positive differential expression
 versus all other clusters. The results are stored as a list of dataframes.
 Each list element contains a named list of clusters (cluster A). Each of
 those list elements contains a dataframe where variables represent the
 comparison to its nearest neighbouring cluster (cluster B) and each sample
 is a gene. There are three variables, named as follows: \code{vs.B.dDR} is
 the difference in detection rate of that gene between the two clusters
 (DR[A] - DR[B]). \code{vs.B.logGER} is the log gene expression ratio
 calculated by taking the difference in mean expression of the gene (see
 \link{meanLogX} for mean calculation) between the two clusters (MTC[A] -
 MTC[B]). \code{vs.B.qVal} is the false discovery rate-corrected p-value of
 the Wilcoxon rank sum test.
}
\description{
Identifies DE genes between nearest neighbour clusters using
\code{\link{calcDEvsCombn}} output for DE results and \code{\link{deDist}} output
to determine nearest neighbours. You probably don't need to use this unless
you're trying to customize \code{\link{clusterWiseDEtest}}.
}
\seealso{
\code{\link{calcDEvsCombn}} for the preceding calculation, and
 \code{\link{clusterWiseDEtest}} for the wrapper function for all DE testing.
}
