% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{CalcSilhouette}
\alias{CalcSilhouette}
\alias{CalcSilhouette,sCVdata-method}
\title{Calculates silhouette widths for all cells in sCVdata}
\usage{
CalcSilhouette(sCVd, inD)

\S4method{CalcSilhouette}{sCVdata}(sCVd, inD)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}
}
\value{
A silhouette object containing the silhouette widths for all cells in
  the sCVdata object.
}
\description{
Calls \code{\link[cluster]{silhouette}} to calculate silhouette widths (a
metric indicating each cell's contribution to cluster cohesion / separation)
for all cells in the sCVdata object, using the cell embedding used in
clustering (see \code{Param(sCVd,"DRforClust")}).
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Calculate silhouette widths for sCVdata
}}

\examples{
\dontrun{
Silhouette(your_sCV_obj) <- CalcSilhouette(sCVd=your_sCV_obj,
                                           inD=your_scRNAseq_data_object)
}

}
\seealso{
\code{\link{CalcSCV}} for wrapper function to calculate all
  statistics for an sCVdata object, \code{\link{fx_calcSilhouette}} for the
  internal function this method points to, and
  \code{\link[cluster]{silhouette}} for the function doing the calculations.
}
