% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{clusterWiseDEtest}
\alias{clusterWiseDEtest}
\title{Cluster-wise differential expression testing}
\usage{
clusterWiseDEtest(il, testAll = TRUE, exponent = 2, pseudocount = 1,
  FDR = 0.01, threshType = "dDR", dDRthresh = 0.15,
  logGERthresh = 1)
}
\arguments{
\item{il}{The list outputted by one of the importData functions (either 
\code{\link{readFromSeurat}} or \code{\link{readFromManual}}).}

\item{exponent}{The log base of your normalized input data. Seurat normalization uses 
the natural log (set this to exp(1)), while other normalization methods generally use 
log2 (set this to 2).}

\item{pseudocount}{The pseudocount added to all log-normalized values in your input 
data. Most methods use a pseudocount of 1 to eliminate log(0) errors.}

\item{FDR}{The false discovery rate to use as a threshold for determining statistical 
significance of differential expression calculated by the Wilcoxon rank-sum test.}

\item{threshType}{Filtering genes for use in differential expression testing can be 
done multiple ways. We use an expression ratio filter for comparing each cluster to 
the rest of the tissue as a whole, but find that difference in detection rates works 
better when comparing clusters to each other. You can set threshType to 
\code{"logGER"} to use a gene expression ratio for all gene filtering, or leave it as 
default (\code{"dDR"}) to use difference in detection rate as the thresholding method 
when comparing clusters to each other.}

\item{dDRthresh}{Magnitude of detection rate difference of a gene between clusters to 
use as filter for determining which genes to test for differential expression between 
clusters.}

\item{logGERthresh}{Magnitude of gene expression ratio for a gene between clusters to 
use as filter for determining which genes to test for differential expression between 
clusters.}
}
\value{
The function returns a list containing the results of differential expression 
  testing for all sets of cluster solutions. This output is passed to the Shiny app. 
  The list contains the following elements:
  \describe{
    \item{CGS}{} 
    \item{deTissue}{} 
    \item{deVS}{}
    \item{deMarker}{}
    \item{deDist}{} 
    \item{deNeighb}{} 
  }
}
\description{
Performs differential expression testing between clusters for all cluster solutions in 
order to assess the biological relevance of each cluster solution. Differential 
expression testing is done using the Wilcoxon rank-sum test implemented in the base R 
\code{stats} package. For details about what is being compared in the tests, see the 
"Value" section.
}
\examples{
\dontrun{
 data_for_scClustViz <- readFromSeurat(your_seurat_object,
                                       convertGeneIDs=F)
 rm(your_seurat_object) 
 # All the data scClustViz needs is in 'data_for_scClustViz'.
 
 DE_for_scClustViz <- clusterWiseDEtest(data_for_scClustViz)
 
 save(data_for_scClustViz,DE_for_scClustViz,
      file="for_scClustViz.RData")
 # Save these objects so you'll never have to run this slow function again!
 
 runShiny(il=data_for_scClustViz,dl=DE_for_scClustViz)
}

}
\seealso{
\code{\link{readFromSeurat}} or \code{\link{readFromManual}} for reading in 
  data to generate the input object for this function, and \code{\link{runShiny}} to 
  use the interactive Shiny GUI to view the results of this testing.
}
