% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{ColNNZ}
\alias{ColNNZ}
\alias{ColNNZ,matrix-method}
\alias{ColNNZ,dgCMatrix-method}
\alias{ColNNZ,dgRMatrix-method}
\alias{ColNNZ,dgTMatrix-method}
\title{Count non-zero entries per column of a matrix}
\usage{
ColNNZ(x)

\S4method{ColNNZ}{matrix}(x)

\S4method{ColNNZ}{dgCMatrix}(x)

\S4method{ColNNZ}{dgRMatrix}(x)

\S4method{ColNNZ}{dgTMatrix}(x)
}
\arguments{
\item{x}{A matrix, hopefully sparse.}
}
\value{
An integer vector of length ncol(x) with the number of non-zero 
  entries in each column of the matrix.
}
\description{
Count non-zero entries per column of a matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: x of class "matrix"

\item \code{dgCMatrix}: x of class "dgCMatrix"

\item \code{dgRMatrix}: x of class "dgRMatrix"

\item \code{dgTMatrix}: x of class "dgTMatrix"
}}

\references{
https://stackoverflow.com/questions/51560456/r-package-matrix-get-number-of-non-zero-entries-per-rows-columns-of-a-sparse
}
