% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{cosineSim}
\alias{cosineSim}
\title{Computes the cosine similarity between two vectors}
\usage{
cosineSim(A, B)
}
\arguments{
\item{A}{One vector in the pair to compare.}

\item{B}{The other vector in the pair to compare.}
}
\value{
A numeric value, the cosine similarity between vectors \code{A} and
  \code{B}
}
\description{
Computes the cosine similarity between two vectors.
}
\references{
\url{https://stats.stackexchange.com/q/31573}/
}
