% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{DEdist}
\alias{DEdist}
\alias{DEdist,sCVdata,numeric-method}
\alias{DEdist,sCVdata,missing-method}
\alias{DEdist,sCVdata,character-method}
\alias{DEdist,sCVdata,matrix-method}
\title{Calculate inter-cluster distances}
\usage{
DEdist(sCVd, y)

\S4method{DEdist}{sCVdata,numeric}(sCVd, y)

\S4method{DEdist}{sCVdata,missing}(sCVd, y)

\S4method{DEdist}{sCVdata,character}(sCVd, y)

\S4method{DEdist}{sCVdata,matrix}(sCVd, y)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{y}{One of the following, depending on the desired distance metric:
\describe{
  \item{By number of differentially expressed genes:}{A numeric vector of 
    length one representing the threshold of false discovery rate 
    controlled for when counting significantly differentially expressed 
    genes.}
  \item{By differential expression significance score:}{Nothing. Only pass
    an argument to \code{sCVd}.}
  \item{By gene expression summary statistic:}{A character vector of length 
    one referring to a column name in \code{ClustGeneStats(sCVd)}: 
    \code{MGE}, \code{DR}, or \code{MDGE}.}
  \item{By distance between centroids in a cell embedding:}{A matrix 
    representing the cell embedding, where cells are rows and dimensions
    are columns. See \code{\link{getEmb}} to easily extract this from your
    input data object.}
}}
}
\value{
A matrix of distances between clusters for each cluster resolution.
  Interpretable by \code{\link[stats]{as.dist}} to generate a \code{dist}
  object.
}
\description{
A variety of methods for calculating inter-cluster distances for sCVdata
objects, most of which return matrices which can be converted to
\code{\link[stats]{dist}} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVd = sCVdata,y = numeric}: By number of differentially-expressed genes (see
\code{\link{fx_calcDist_numDE}})

\item \code{sCVd = sCVdata,y = missing}: By differential expression significance score (see
\code{\link{fx_calcDist_scoreDE}})

\item \code{sCVd = sCVdata,y = character}: By gene expression summary statistic (see
\code{\link{fx_calcDist_geneExpr}})

\item \code{sCVd = sCVdata,y = matrix}: By distance between centroids in a cell embedding (see
\code{\link{fx_calcDist_cellEmb}})
}}

\examples{
\dontrun{
numDE_bw_clusts <- DEdist(sCVd=your_sCV_obj,0.01)
DEdistNN(numDE_bw_clusts)
}

}
\seealso{
\code{\link{DEdistNN}} for a method to find nearest neighbouring
  clusters from \code{DEdist} results.
}
