% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{DEdistNN}
\alias{DEdistNN}
\alias{DEdistNN,numeric-method}
\alias{DEdistNN,matrix-method}
\title{Determine nearest neighbouring clusters}
\usage{
DEdistNN(x)

\S4method{DEdistNN}{numeric}(x)

\S4method{DEdistNN}{matrix}(x)
}
\arguments{
\item{d}{Output of \code{\link{DEdist}}}
}
\value{
A named integer vector where each name is a cluster and element is
  the position of its nearest neighbouring cluster in \code{Clusters(sCVd)}.
}
\description{
Finds nearest neighbouring clusters from output of \code{\link{DEdist}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Nearest neighbours from DEdist vector.

\item \code{matrix}: Nearest neighbours from DEdist matrix.
}}

\examples{
\dontrun{
numDE_bw_clusts <- DEdist(sCVd=your_sCV_obj,0.01)
DEdistNN(numDE_bw_clusts)
}

}
