% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{DEmarker}
\alias{DEmarker}
\alias{DEmarker,sCVdata-method}
\title{Find marker genes for each cluster}
\usage{
DEmarker(sCVd, FDRthresh)

\S4method{DEmarker}{sCVdata}(sCVd, FDRthresh)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of named numeric vectors, one entry for each cluster.
  Each vector is the FDR values for each gene that is significantly
  positively differentially expressed in that cluster over all other clusters
  in pairwise tests, with names corresponding to gene labels.
}
\description{
Identifies the significantly positively differentially expressed genes
between a cluster and all other clusters, using the pairwise differential
gene expression testing results.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Find marker genes from sCVdata
}}

