% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{DEneighb}
\alias{DEneighb}
\alias{DEneighb,sCVdata-method}
\title{Find differentially expressed genes between nearest neighbouring clusters}
\usage{
DEneighb(sCVd, FDRthresh)

\S4method{DEneighb}{sCVdata}(sCVd, FDRthresh)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of named numeric vectors, one entry for each nearest
  neighbouring pair. Each vector is the FDR values for each gene that is
  significantly differentially expressed between that pair of clusters, with
  names corresponding to gene labels.
}
\description{
Identifies the significantly differentially expressed genes between nearest
neighbouring clusters, using the pairwise differential gene expression
testing results.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Find DE genes between nearest neighbours from sCVdata
}}

