% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{findKeyType}
\alias{findKeyType}
\title{Automatically determines keytype for AnnotationDb lookup}
\usage{
findKeyType(nge, annotationDB)
}
\arguments{
\item{nge}{The gene expression matrix, i.e. \code{getExpr(yourDataObject)}.}

\item{annotationDB}{The AnnotationDb object.}
}
\value{
A character value, the \code{keytype} of the AnnotationDb object, to
  be passed to \code{map2symbol}.
}
\description{
Compares rownames of gene expression matrix with keys for all keytypes in an
AnnotationDb object, and returns the keytype with the best match.
}
\details{
If the rownames for the gene expression matrix are not official gene symbols
(determined by less than 80 percent of rownames matching keys in the SYMBOL keytype
of the AnnotationDb object), this function compares rownames of gene
expression matrix with keys for all keytypes in an AnnotationDb object, and
returns the keytype with the best match.
}
\seealso{
\code{\link{map2symbol}}
}
