% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcCGS}
\alias{fx_calcCGS}
\title{Internal fx for cluster-wise gene statistics}
\usage{
fx_calcCGS(nge, cl, exponent, pseudocount)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{exponent}{The log base of your normalized input data. Seurat
normalization uses the natural log (set this to exp(1)), while other
normalization methods generally use log2 (set this to 2).}

\item{pseudocount}{The pseudocount added to all log-normalized values in your
input data. Most methods use a pseudocount of 1 to eliminate log(0) errors.}
}
\value{
The function returns a list of dataframes. Each list element contains
  a named list of clusters at that resolution. Each of those list elements
  contains a dataframe of three variables, where each sample is a gene.
  \code{DR} is the proportion of cells in the cluster in which that gene was
  detected. \code{MDGE} is mean normalized gene expression for that gene in
  only the cells in which it was detected (see \code{\link{meanLogX}} for
  mean calculation). \code{MGE} is the mean normalized gene expression for
  that gene in all cells of the cluster (see \code{\link{meanLogX}} for mean
  calculation).
}
\description{
Internal function. See \code{\link{CalcCGS}}.
}
