% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcCGS_BP}
\alias{fx_calcCGS_BP}
\title{Internal fx for cluster-wise gene statistics using BiocParallel}
\usage{
fx_calcCGS_BP(nge, cl, exponent, pseudocount)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{exponent}{A length-one numeric vector representing the
base of the log-normalized gene expression data to be processed. Generally
gene expression data is transformed into log2 space when normalizing (set
this to 2), though \code{Seurat} uses the natural log (set this to exp(1)). 
If you are using data that has not been log-transformed (for example, 
corrected counts from SCTransform), set this to NA.}

\item{pseudocount}{A length-one numeric vector representing the
pseudocount added to all log-normalized values in your input data. Most
methods use a pseudocount of 1 to eliminate log(0) errors. If you are using 
data that has not been log-transformed (for example, corrected counts from 
SCTransform), set this to NA.}
}
\value{
The function returns a list of dataframes. Each list element contains
  a named list of clusters at that resolution. Each of those list elements
  contains a dataframe of three variables, where each sample is a gene.
  \code{DR} is the proportion of cells in the cluster in which that gene was
  detected. \code{MDGE} is mean normalized gene expression for that gene in
  only the cells in which it was detected (see \code{\link{meanLogX}} for
  mean calculation). \code{MGE} is the mean normalized gene expression for
  that gene in all cells of the cluster (see \code{\link{meanLogX}} for mean
  calculation).
}
\description{
Internal function. See \code{\link{CalcCGS}}.
}
