% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDEcombn_BP}
\alias{fx_calcDEcombn_BP}
\title{Internal fx to calculate DE between combinations of clusters using BiocParallel}
\usage{
fx_calcDEcombn_BP(nge, cl, deMes)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{deMes}{The output from \code{\link{fx_calcEScombn}}.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
  gene in each cluster to that gene in every other cluster in a series of
  tests. The results are stored as a nested list of dataframes. Each list
  element contains a named list of clusters (cluster A). Each of those lists
  contains a named list of all the other clusters (cluster B). Each of those
  list elements contains a dataframe of four variables, where each sample is
  a gene. \code{dDR} is the difference in detection rate of that gene between
  the two clusters (DR[A] - DR[B]). \code{logGER} is the log gene expression
  ratio calculated by taking the difference in mean expression of the gene
  (see \code{\link{meanLogX}} for mean calculation) between the two clusters
  (MGE[A] - MGE[B]). \code{Wstat} and \code{pVal} are the test statistic and
  the p-value of the Wilcoxon rank sum test. \code{FDR} is the false
  discovery rate-corrected p-value of the test.
}
\description{
Internal function. See \code{\link{CalcDEcombn}}.
}
\details{
Calculates Wilcoxon rank-sum tests for all genes in each of the potential
combinations of clusters to compare.
}
