% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDEvsRest_BP}
\alias{fx_calcDEvsRest_BP}
\title{Internal fx to perform one vs all DE testing using BiocParallel}
\usage{
fx_calcDEvsRest_BP(nge, cl, deTes)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{deTes}{The output from \code{\link{fx_calcESvsRest}}.}
}
\value{
Differential testing results from Wilcoxon rank sum tests comparing a
  gene in each cluster to the rest of the cells as a whole in a one vs all
  comparison. The results are stored as a named list of dataframes. There is
  a list element for each cluster containing a dataframe of three variables,
  where each sample is a gene. \code{logGER} is the log gene expression ratio
  calculated by subtracting the mean expression of the gene (see
  \link{meanLogX} for mean calculation) in all other cells from the mean
  expression of the gene in this cluster. \code{Wstat} and \code{pVal} are
  the test statistic and the p-value of the Wilcoxon rank sum test.
  \code{FDR} is the false discovery rate-corrected p-value of the test.
}
\description{
Internal function. See \code{\link{CalcDEvsRest}}.
}
\details{
Calculates Wilcoxon rank-sum tests for all genes in each one-vs-all
comparison of a cluster vs the rest of the data. You probably don't need to
use this unless you're trying to customize \code{\link{clusterWiseDEtest}}.
}
