% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDist_cellEmb}
\alias{fx_calcDist_cellEmb}
\title{Internal function for distance between clusters from cell embedding}
\usage{
fx_calcDist_cellEmb(cl, cellEmb)
}
\arguments{
\item{cellEmb}{A cell embedding with cells in rows and dimensions in columns.}

\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}}
}
\value{
A matrix of distances between clusters for each cluster resolution.
  Interpretable by \code{\link[stats]{as.dist}} to generate a \code{dist}
  object.
}
\description{
Internal function. See \code{\link{DEdist}}.
}
\details{
Calculates euclidean distance between centroids of each cluster in the cell
embedding used for clustering. Calculated using \code{\link[stats]{dist}}.
}
