% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDist_geneExpr}
\alias{fx_calcDist_geneExpr}
\title{Internal function for distance between clusters by gene expression statistic.}
\usage{
fx_calcDist_geneExpr(CGS, metric)
}
\arguments{
\item{metric}{A summary statistic of gene expression per cluster, from
\code{\link{CalcCGS}}}

\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}}
}
\value{
A matrix of distances between clusters for each cluster resolution.
  Interpretable by \code{\link[stats]{as.dist}} to generate a \code{dist}
  object.
}
\description{
Internal function. See \code{\link{DEdist}}.
}
\details{
Calculates euclidean distance in gene expression space between each pair of
clusters, using a summary statistic of gene expression for each cluster.
Calculated using \code{\link[stats]{dist}}.
}
