% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDist_numDE}
\alias{fx_calcDist_numDE}
\title{Internal function for distance between clusters by number of DE genes.}
\usage{
fx_calcDist_numDE(deVS, FDRthresh)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A matrix of distances between clusters for each cluster resolution.
  Interpretable by \code{\link[stats]{as.dist}} to generate a \code{dist}
  object.
}
\description{
Internal function. See \code{\link{DEdist}}.
}
\details{
Counts number of DE genes between each pair of clusters from
\code{\link{CalcDEcombn}} output.
}
