% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcDist_scoreDE}
\alias{fx_calcDist_scoreDE}
\title{Internal function for distance between clusters by DE score.}
\usage{
fx_calcDist_scoreDE(deVS)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}}
}
\value{
A matrix of distances between clusters for each cluster resolution.
  Interpretable by \code{\link[stats]{as.dist}} to generate a \code{dist}
  object.
}
\description{
Internal function. See \code{\link{DEdist}}.
}
\details{
Sums absolute DE scores between each pair of clusters from
\code{\link{CalcDEcombn}} output. DE scores are \code{-log10(FDR)}. Used to
determine nearest neighbouring clusters.
}
