% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcEScombn}
\alias{fx_calcEScombn}
\title{Internal fx to calculate genes used for deVS calculation}
\usage{
fx_calcEScombn(cl, CGS, DRthresh)
}
\arguments{
\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{CGS}{The output from \code{\link{fx_calcCGS}}.}

\item{DRthresh}{The threshold for minimum detection rate of a gene to be
considered in the following Wilcoxon rank-sum test. Gene must pass
threshold in at least one of the pair of clusters.}
}
\value{
The function returns a list where each list element is a dataframe
  with effect size statistics (log-ratios of gene expression and difference
  in detection rate) when comparing each gene in a cluster to the rest of the
  cells as a whole in a one vs all comparison. These dataframes are filtered
  to only include those gene that pass the relevant threshold, and thus the
  rownames for each list entry correspond to the genes to test in
  \code{\link{fx_calcDEcombn}}.
}
\description{
Internal function. See \code{\link{CalcDEcombn}}.
}
\details{
Calculates the log-ratios of gene expression and difference in detection rate
for all genes in each of the potential combinations of clusters to compare.
This is used to determine the genes used in deVS calculations.
}
