% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcESvsRest}
\alias{fx_calcESvsRest}
\title{Internal fx to calculate logGER for DEvsRest calculation}
\usage{
fx_calcESvsRest(nge, cl, CGS, exponent, pseudocount, DRthresh)
}
\arguments{
\item{nge}{The log-normalized gene expression matrix.}

\item{cl}{The factor with cluster assignments per cell (column of nge).}

\item{CGS}{The output from \code{\link{CalcCGS}}.}

\item{exponent}{The log base of your normalized input data. Seurat
normalization uses the natural log (set this to exp(1)), while other
normalization methods generally use log2 (set this to 2).}

\item{pseudocount}{The pseudocount added to all log-normalized values in your
input data. Most methods use a pseudocount of 1 to eliminate log(0) errors.}

\item{DRthresh}{The threshold for minimum detection rate of a gene in the
cluster for the gene to be considered in the following Wilcoxon rank-sum
test.}
}
\value{
The function returns a list where each list element is the log-ratios
  of gene expression when comparing each gene in a cluster to the rest of the
  cells as a whole in a one vs all comparison. These logGER tables are
  filtered to only include those gene that pass logGER threshold, and thus
  the names for each list entry correspond to the genes to test in
  \code{\link{fx_calcDEvsRest}}.
}
\description{
Internal function. See \code{\link{CalcDEvsRest}}.
}
\details{
Calculates the log-ratios of gene expression for all genes in each one-vs-all
comparison of a cluster vs the rest of the data. This is used to determine
the genes used in DEvsRest calculations.
}
