% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcMarker}
\alias{fx_calcMarker}
\title{Internal function to find marker genes for each cluster}
\usage{
fx_calcMarker(deVS, FDRthresh)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of data frames, one entry for each cluster. Each data
  frame is the combined results of the set of pairwise gene expression tests
  comparing the cluster to each other cluster in the data. See
  \code{\link{CalcDEcombn}} and \code{\link{DEcombn}} for details.
}
\description{
Internal function. See \code{\link{DEmarker}}.
}
\details{
Identifies the significantly positively differentially expressed genes
between a cluster and all other clusters, using the pairwise differential
gene expression testing results.
}
