% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcNeighb}
\alias{fx_calcNeighb}
\title{Internal function to find DE genes between nearest neighbouring clusters}
\usage{
fx_calcNeighb(deVS, NN, FDRthresh)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}.}

\item{NN}{A vector of nearest neighbours from \code{\link{DEdistNN}}.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of data frames, one entry for each cluster. Each data
  frame is the results of the gene expression test comparing the cluster to
  its nearest neighbour. See \code{\link{CalcDEcombn}} and
  \code{\link{DEcombn}} for details.
}
\description{
Internal function. See \code{\link{DEneighb}}.
}
\details{
Identifies the significantly differentially expressed genes between nearest
neighbouring clusters, using the pairwise differential gene expression
testing results.
}
