% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcSilhouette}
\alias{fx_calcSilhouette}
\title{Internal fx to call \code{\link[cluster]{silhouette}} for sCVdata.}
\usage{
fx_calcSilhouette(pca, cl)
}
\arguments{
\item{pca}{The cell embeddings used to calculate clusters.}

\item{cl}{The cluster assignments.}
}
\value{
A silhouette object.
}
\description{
Internal function. See \code{\link{CalcSilhouette}}.
}
