% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccess.R
\name{getExpr}
\alias{getExpr}
\title{Get gene expression matrix from input data object}
\usage{
getExpr(x, assayType)
}
\arguments{
\item{x}{The single-cell data object.}

\item{assayType}{The name of the assay slot to access. Not required for 
Seurat v1 or v2.}
}
\description{
Extract the gene expression matrix from a single-cell data object containing
the input data for scClustViz visualization.
}
\details{
This is a wrapper function to the relevant class's normalized data slot
accessor method. Currently supported input object classes:
\itemize{
  \item Class \code{\link[Seurat]{seurat}/\link[Seurat]{Seurat}} stored in
    \code{x@data} or \code{x@assays[[assayType]]@data}, 
    depending on Seurat object version.
  \item Class \code{\link[SingleCellExperiment]{SingleCellExperiment}} 
    accessed by \code{\link[SummarizedExperiment]{assay}(x,assayType)}.
}
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}
}
