% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccess.R
\name{getMD}
\alias{getMD}
\title{Get metadata from input data object}
\usage{
getMD(x)
}
\arguments{
\item{x}{The single-cell data object.}
}
\description{
Extract the cell metadata data frame from a single-cell data object
containing the input data for scClustViz visualization.
}
\details{
This is a wrapper function to the relevant class's cell metadata slot
accessor / assignment method. Currently supported input object classes:
\itemize{
  \item Class \code{\link[Seurat]{seurat}/\link[Seurat]{Seurat}} accessed by 
    \code{x@data.info} or \code{x@meta.data}, 
    depending on Seurat object version.
  \item Class \code{\link[SingleCellExperiment]{SingleCellExperiment}}
    accessed by \code{\link[SingleCellExperiment]{colData}(x)}.
}
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}
}
