% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{map2symbol}
\alias{map2symbol}
\title{Maps gene identifiers to official gene symbol}
\usage{
map2symbol(nge, annotationDB, rownameKeytype)
}
\arguments{
\item{nge}{The gene expression matrix, i.e. \code{getExpr(yourDataObject)}.}

\item{annotationDB}{The AnnotationDb object.}

\item{rownameKeytype}{The keytype of the gene identifiers. Can be determined
automatically using \code{findKeyType}.}
}
\value{
A named character vector, the output of \code{AnnotationDbi::mapIds}
  where values are gene symbols and names are the input gene identifiers.
}
\description{
Maps gene identifiers in the gene expression matrix to official gene symbols
for ease of data interpretation in scClustViz UI.
}
\details{
If rownames are already official gene symbols, returns null.  In this case,
\code{addCellMarkersToCGS} and \code{labelCellTypes} will proceed under the
assumption that the gene identifiers in the input data are already official
gene symbols.
}
\seealso{
\code{\link{findKeyType}},\code{\link[AnnotationDbi]{mapIds}},
  \code{\link{addCellMarkersToCGS}}, and \code{\link{labelCellTypes}}.
}
