% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadLabels2.R
\name{origin.label.dist}
\alias{origin.label.dist}
\title{Internal fx to calc dist from point to label end}
\usage{
origin.label.dist(i, ld, oX, oY)
}
\arguments{
\item{i}{index of column of \code{ld}}

\item{ld}{labelData - rows are xy coords and xy sizes, columns are labels}

\item{oX}{original x coord}

\item{oY}{original y coord}
}
\description{
Internal fx to calc dist from point to label end
}
