% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadLabels2.R
\name{overlap.edge}
\alias{overlap.edge}
\title{Internal fx to calc overlap of label and plot edge}
\usage{
overlap.edge(i, ld)
}
\arguments{
\item{i}{index of column of \code{ld}}

\item{ld}{labelData - rows are xy coords and xy sizes, columns are labels}
}
\description{
Internal fx to calc overlap of label and plot edge
}
