% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadLabels.R
\name{overlap.proportion.point}
\alias{overlap.proportion.point}
\title{Internal fx to calc relative overlap of labels}
\usage{
overlap.proportion.point(i, ld, oX, oY)
}
\arguments{
\item{i}{index of column of \code{ld}}

\item{ld}{labelData - rows are xy coords and xy sizes, columns are labels}

\item{oX}{original x coord}

\item{oY}{original y coord}
}
\description{
Internal fx to calc relative overlap of labels
}
