% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_clusterGenes_DEgenes}
\alias{plot_clusterGenes_DEgenes}
\title{scClustViz plot: Plot within-cluster gene expression highlighting DE genes}
\usage{
plot_clusterGenes_DEgenes(sCVd, selClust, DEgenes, DEnum, DEtype)
}
\arguments{
\item{sCVd}{The sCVdata object.}

\item{selClust}{A named character vector representing the cluster to be
displayed. If \code{\link{labelCellTypes}} has been run, pass the
appropriate element of \code{attr(Clusters(sCV),"ClusterNames")} to this
argument to show both cluster number and cell type label in the legend.}

\item{DEgenes}{The output of \code{\link{dotplotDEgenes}}.}

\item{DEnum}{Single integer representing the maximum number of DE genes per
cluster to include in the plot.}

\item{DEtype}{One of: \code{"DEvsRest"} - see \code{\link{DEvsRest}};
\code{"DEneighb"} - see \code{\link{DEneighb}}; \code{"DEmarker"} - see
\code{\link{DEmarker}}.}
}
\description{
This function makes a scatterplot of gene detection rate vs. mean detected
gene abundance, highlighting differentially expressed genes.
}
\examples{
\dontrun{
pdf("filepath.pdf",width=12,height=7)
plot_clusterGenes_DEgenes(sCVd=sCVdata,
                          selClust="1",
                          DEgenes=dotplotDEgenes(sCVdata,
                                                 DEtype="DEneighb",
                                                 FDRthresh=0.01),
                          DEnum=5,
                          DEtype="DEneighb")
dev.off()
}

}
