% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_clusterGenes_markers}
\alias{plot_clusterGenes_markers}
\title{scClustViz plot: Plot within-cluster gene expression highlighting marker
genes}
\usage{
plot_clusterGenes_markers(sCVd, selClust, cellMarkers)
}
\arguments{
\item{sCVd}{The sCVdata object.}

\item{selClust}{A named character vector representing the cluster to be
displayed. If \code{\link{labelCellTypes}} has been run, pass the
appropriate element of \code{attr(Clusters(sCV),"ClusterNames")} to this
argument to show both cluster number and cell type label in the legend.}

\item{cellMarkers}{The \code{cellMarkers} argument from
\code{\link{runShiny}}. A list of marker genes for expected cell types.}
}
\description{
This function makes a scatterplot of gene detection rate vs. mean detected
gene abundance, highlighting genes identified as cell type specific markers
by the user. \strong{This function will not work unless
\code{\link{addCellMarkersToCGS}} has been run on the sCVdata object prior.}
}
\examples{
\dontrun{
cellMarkers <- list("Cortical precursors"=c("Mki67","Sox2","Pax6",
                                                  "Pcna","Nes","Cux1","Cux2"),
                          "Interneurons"=c("Gad1","Gad2","Npy","Sst","Lhx6",
                                           "Tubb3","Rbfox3","Dcx"),
                          "Cajal-Retzius neurons"="Reln",
                          "Intermediate progenitors"="Eomes",
                          "Projection neurons"=c("Tbr1","Satb2","Fezf2",
                                                 "Bcl11b","Tle4","Nes",
                                                 "Cux1","Cux2","Tubb3",
                                                 "Rbfox3","Dcx")
                          )
sCVdata <- labelCellTypes(sCVdata,
                          cellMarkers=cellMarkers,
                          symbolMap=NULL)

pdf("filepath.pdf",width=12,height=7)
plot_clusterGenes_markers(sCVd=sCVdata,
                          selClust="1",
                          cellMarkers=cellMarkers)
dev.off()
}

}
