% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_clustSep}
\alias{plot_clustSep}
\title{scClustViz plot: Cluster separation boxplots}
\usage{
plot_clustSep(sCVdL, DEtype, FDRthresh = 0.05, res, Xlim, Ylim)
}
\arguments{
\item{sCVdL}{A named list of sCVdata objects, output of
\code{\link{CalcAllSCV}}.}

\item{DEtype}{One of "DEneighb", "DEmarker", or "silWidth". "DEneighb" shows
number of significantly differentially expressed genes between nearest
neighbouring clusters. "DEmarker" shows number of marker genes per cluster,
significantly positively differentially expressed genes in all pairwise
comparisons with other clusters. "silWidth" shows silhouette widths with
average silhouette width as a trace across all clustering solutions. (see
\code{\link[cluster]{silhouette}}).}

\item{FDRthresh}{Default=0.05. The false discovery rate threshold for
determining significance of differential gene expression.}

\item{res}{Optional. Name of cluster resolution to highlight. Must be one of
\code{names(sCVdL)}.}

\item{Xlim}{Optional. Passed to
\code{\link[graphics]{plot.default}(xlim=Xlim)}.}

\item{Ylim}{Optional. Passed to
\code{\link[graphics]{plot.default}(ylim=Ylim)}.}
}
\description{
This function plots metrics of cluster solution cohesion or overfitting as a
function of the number of clusters found.
}
\examples{
\dontrun{
plot_clustSep(sCVdL,DEtype="DEneighb",FDRthresh=0.05,res="res.0.8")
}

}
