% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_compareClusts}
\alias{plot_compareClusts}
\title{scClustViz plot: Volcano and MA-style plots to compare clusters}
\usage{
plot_compareClusts(sCVd, clA, clB, dataType, labType = "de", labGenes,
  labNum = 5, labTypeDiff = "logGER")
}
\arguments{
\item{sCVd}{The sCVdata object.}

\item{clA}{Cluster identifier for side A of the comparison.}

\item{clB}{Cluster identifier for side B of the comparison.}

\item{dataType}{For MA-style plots comparing difference and mean of gene
summary statistics, one of: \code{"DR"} (detection rate); \code{"MGE"}
(mean gene expression); \code{"MDGE"} (mean detected gene expression). For
volcano plots, the effect size measure can be one of: \code{"dDR"}
(difference in detection rate); \code{"logGER"} (log gene expression
ratio). To compare relationship between difference in detection rate and
log gene expression ratio, use \code{"GERvDDR"}.}

\item{labType}{Default="de". A character vector indicating which genes to
highlight. One of \code{"de"} (most statistically significant genes),
\code{"diff"} (most different by dataType shown), or \code{"search"}
(specified genes).}

\item{labGenes}{Only required if \code{labType="search"}. Gene names to
highlight.}

\item{labNum}{Default=5. Number of genes to highlight per side.}

\item{labTypeDiff}{Default="logGER". Only required if
\code{dataType="GERvDDR"} and \code{labType="diff"}. Which axis to use for
difference calculation. One of \code{"dDR"} (difference in detection rate)
or \code{"logGER"} (log gene expression ratio).}
}
\description{
This function generates scatterplots inspired by volcano and MA plots for
comparing gene expression between pairs of clusters.
}
\examples{
\dontrun{
plot_compareClusts(sCVdata,
                   clA="1",
                   clB="2",
                   dataType="GERvDDR",
                   labType="search",
                   labGenes="Actb")
}

}
