% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_deDotplot}
\alias{plot_deDotplot}
\title{scClustViz plot: Plot gene expression dotplots.}
\usage{
plot_deDotplot(sCVd, DEgenes, DEnum)
}
\arguments{
\item{sCVd}{The sCVdata object.}

\item{DEgenes}{The output of \code{\link{dotplotDEgenes}}.}

\item{DEnum}{Single integer representing the maximum number of DE genes per
cluster to include in the dotplot.}
}
\description{
This function makes dotplots (a heatmap analogue) showing gene expression for
a set of genes across all clusters.
}
\details{
When generated in an interactive context (i.e. RStudio), this can sometimes
result in a \code{figure margins too large} error. See example for suggested
dimensions of the graphic device.
}
\examples{
\dontrun{
pdf("filepath.pdf",width=11,height=7)
plot_deDotplot(sCVd=sCVdata,
               DEgenes=dotplotDEgenes(sCVdata,
                                      DEtype="DEneighb",
                                      FDRthresh=0.01)
               DEnum=5)
dev.off()
}

}
