% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_GEboxplot}
\alias{plot_GEboxplot}
\title{scClustViz plot: Compare gene expression across clusters}
\usage{
plot_GEboxplot(nge, sCVd, gene, geneName, opts = c("sct", "dr"))
}
\arguments{
\item{nge}{The gene expression matrix, see \code{\link{getExprs}}.}

\item{sCVd}{The sCVdata object.}

\item{gene}{The gene to display.}

\item{geneName}{Optional. A named character vector of length one. The element
is the full gene name, and the name is the gene symbol.}

\item{opts}{Default=\code{c("sct","dr")}. A character vector with plotting
options. If it includes \code{"sct"}, data points will be overlaid as a
jitter over the boxplot. If it includes \code{"dr"}, detection rate per
cluster will be plotted as a small black bar over each boxplot, with the
corresponding axis on the right.}
}
\description{
This function generates boxplots comparing normalized gene abundance across
all clusters.
}
\examples{
\dontrun{
plot_GEboxplot(getExpr(input_data_obj),
               sCVd=sCVdata,
               gene="Actb")
}

}
