% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_mdCompare}
\alias{plot_mdCompare}
\title{scClustViz plot: Plot to compare cell metadata}
\usage{
plot_mdCompare(MD, mdX, mdY, sel_cells, sel_clust, md_log)
}
\arguments{
\item{MD}{A dataframe of cellular metadata. See \code{\link{getMD}}.}

\item{mdX}{A character vector of one refering to the variable name from
\code{MD} to plot on the x-axis.}

\item{mdY}{A character vector of one refering to the variable name from
\code{MD} to plot on the y-axis.}

\item{sel_cells}{Optional. A character vector of cell names (rownames of
\code{MD}) to highlight in the plot.}

\item{sel_clust}{Optional. The name of the selected cluster
(\code{sel_cells}) to include in the legend. If
\code{\link{labelCellTypes}} has been run, pass the appropriate element of
\code{attr(Clusters(sCV),"ClusterNames")} to this argument to show both
cluster number and cell type label in the legend.}

\item{md_log}{Optional. A character vector indicating which axes should be
log scaled. \code{c("x","y")} to log-scale both axes.}
}
\description{
This function makes scatter/boxplots comparing cellular metadata.
}
\examples{
\dontrun{
plot_mdCompare(MD=getMD(input_data_obj),
               mdX="total_counts",
               mdY="total_features",
               sel_cells=names(Clusters(sCVdata))[Clusters(sCVdata) == "1"],
               sel_clust="1",
               md_log="x")
}

}
