% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{plot_sil}
\alias{plot_sil}
\title{scClustViz plot: Silhouette plot}
\usage{
plot_sil(sCVd)
}
\arguments{
\item{sCVd}{An \code{\link{sCVdata}} object with a non-null \code{Silhouette}
slot.}
}
\description{
This function is a wrapper to \code{plot(silhouette(x))}.
}
