% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{rainbow2}
\alias{rainbow2}
\title{Generates a rainbow colour-scale}
\usage{
rainbow2(n)
}
\arguments{
\item{n}{Number of colours to include.}
}
\description{
Generates a slightly desaturated rainbow colour-scale that's a little easier on the
eyes than the R default.
}
