% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{readFromSeurat}
\alias{readFromSeurat}
\title{Read in data from a Seurat object automatically}
\usage{
readFromSeurat(inD)
}
\arguments{
\item{inD}{A Seurat object containing slots as outlined in Details.}
}
\value{
The function returns a list containing input data necessary for both
  the cluster-wise differential expression testing function and the Shiny app
  itself. The list contains the following elements: 
  \describe{ 
    \item{nge}{The normalized gene expression matrix.} 
    \item{md}{The metadata dataframe, not including cluster assignments.} 
    \item{cl}{The cluster assignment dataframe, containing cluster 
      assignments for each resolution tested. The columns will be sorted in 
      order of increasing resolution (k, number of clusters).}
    \item{dr_clust}{The cell embeddings used in the clustering, from PCA.}
    \item{dr_viz}{The cell embeddings used for visualization in 2D, from 
      tSNE.}
  }
}
\description{
Loads the necessary data from a Seurat object for use in both the
cluster-wise differential expression testing function, as well as in the
Shiny app itself.
}
\section{Seurat object slots}{
 The following slots are expected in the Seurat
  object. If you're using Seurat v1.x, the equivalent slots are expected
  (this code takes advantage of \code{UpdateSeuratObject} to find the
  relevant data in older Seurat objects.) 
  \describe{ 
    \item{@data}{Holds the normalized gene expression matrix.} 
    \item{@meta.data}{Holds the metadata, including cluster assignments. 
      \emph{Cluster assignment columns of the metadata should be titled with 
      their resolution parameters, as is the default in Seurat (ie. 
      "res.0.8").} }
    \item{@dr$pca@cell.embeddings}{Holds the results of the PCA run by 
      Seurat. The cell embeddings are used for the silhouette plot in the 
      Shiny app. If Seurat v2.x or greater was used, only the PC dimensions 
      used in clustering will be considered in silhouette calculations. If 
      an alternative dimensionality reduction method was used prior to 
      clustering, use \code{readFromManual} to manually specify the desired 
      cell embeddings.} 
    \item{@dr$tsne@cell.embeddings}{Holds the results of the tSNE run by 
      Seurat. The cell embeddings are used for cell visualizations in the 
      Shiny app. If an alternative 2D projection method was used, use 
      \code{readFromManual} to manually specify the desired cell embeddings.} 
  }
}

\examples{
\dontrun{
 data_for_scClustViz <- readFromSeurat(your_seurat_object)
 rm(your_seurat_object)
 # All the data scClustViz needs is in 'data_for_scClustViz'.

 DE_for_scClustViz <- clusterWiseDEtest(data_for_scClustViz)

 save(data_for_scClustViz,DE_for_scClustViz,
      file="for_scClustViz.RData")
 # Save these objects so you'll never have to run this slow function again!

 runShiny(filePath="for_scClustViz.RData")
}

}
\seealso{
https://satijalab.org/seurat/ for more information on the Seurat
  package, and \code{\link{readFromManual}} for loading data by manually
  passing the requisite data objects.

Other importData functions: \code{\link{readFromManual}}
}
\concept{importData functions}
