% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\docType{methods}
\name{RowNNZ}
\alias{RowNNZ}
\alias{RowNNZ,matrix-method}
\alias{RowNNZ,dgCMatrix-method}
\alias{RowNNZ,dgRMatrix-method}
\alias{RowNNZ,dgTMatrix-method}
\title{Count non-zero entries per row of a matrix}
\usage{
RowNNZ(x)

\S4method{RowNNZ}{matrix}(x)

\S4method{RowNNZ}{dgCMatrix}(x)

\S4method{RowNNZ}{dgRMatrix}(x)

\S4method{RowNNZ}{dgTMatrix}(x)
}
\arguments{
\item{x}{A matrix, hopefully sparse.}
}
\value{
An integer vector of length nrow(x) with the number of non-zero 
  entries in each row of the matrix.
}
\description{
Count non-zero entries per row of a matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: x of class "matrix"

\item \code{dgCMatrix}: x of class "dgCMatrix"

\item \code{dgRMatrix}: x of class "dgRMatrix"

\item \code{dgTMatrix}: x of class "dgTMatrix"
}}

\references{
https://stackoverflow.com/questions/51560456/r-package-matrix-get-number-of-non-zero-entries-per-rows-columns-of-a-sparse
}
