% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCVdataClass.R
\docType{class}
\name{sCVdata-class}
\alias{sCVdata-class}
\alias{sCVdata}
\alias{Clusters,sCVdata-method}
\alias{Clusters}
\alias{Clusters<-,sCVdata-method}
\alias{ClusterNames,sCVdata-method}
\alias{ClusterNames<-,sCVdata-method}
\alias{ClustGeneStats,sCVdata-method}
\alias{ClustGeneStats}
\alias{ClustGeneStats<-,sCVdata-method}
\alias{DEvsRest,sCVdata-method}
\alias{DEvsRest}
\alias{DEvsRest<-,sCVdata-method}
\alias{DEcombn,sCVdata-method}
\alias{DEcombn}
\alias{DEcombn<-,sCVdata-method}
\alias{Silhouette,sCVdata-method}
\alias{Silhouette}
\alias{Silhouette<-,sCVdata-method}
\alias{Param,sCVdata-method}
\alias{Param}
\title{Analysis results for the scClustViz app}
\usage{
\S4method{Clusters}{sCVdata}(sCVd)

\S4method{Clusters}{sCVdata}(sCVd) <- value

\S4method{ClusterNames}{sCVdata}(sCVd)

\S4method{ClusterNames}{sCVdata}(sCVd) <- value

\S4method{ClustGeneStats}{sCVdata}(sCVd)

\S4method{ClustGeneStats}{sCVdata}(sCVd) <- value

\S4method{DEvsRest}{sCVdata}(sCVd)

\S4method{DEvsRest}{sCVdata}(sCVd) <- value

\S4method{DEcombn}{sCVdata}(sCVd)

\S4method{DEcombn}{sCVdata}(sCVd) <- value

\S4method{Silhouette}{sCVdata}(sCVd)

\S4method{Silhouette}{sCVdata}(sCVd) <- value

\S4method{Param}{sCVdata}(sCVd, param)
}
\description{
An S4 class to store analysis results for the scClustViz app. sCVdata objects
should be generated using the function \code{\link{CalcSCV}}, which at
minimum populates the slots \code{Clusters}, \code{ClustGeneStats}, and
\code{params}. The remaining slots can be optionally populated - see slot
entries for details. For efficiency, multiple sCVdata objects can be
generated for multiple cluster solutions using the single command
\code{\link{CalcAllSCV}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{Clusters}: Access Clusters slot

\item \code{Clusters<-}: Assign Clusters slot

\item \code{ClusterNames}: Access Clusters slot

\item \code{ClusterNames<-}: Assign Clusters slot

\item \code{ClustGeneStats}: Access ClustGeneStats slot

\item \code{ClustGeneStats<-}: Assign ClustGeneStats slot

\item \code{DEvsRest}: Access DEvsRest slot

\item \code{DEvsRest<-}: Assign DEvsRest slot - see slot details

\item \code{DEcombn}: Access DEcombn slot

\item \code{DEcombn<-}: Assign DEcombn slot - see slot details

\item \code{Silhouette}: Access Silhouette slot

\item \code{Silhouette<-}: Assign Silhouette slot

\item \code{Param}: Access Param slot (see \code{\link{sCVparams}})
}}

\section{Slots}{

\describe{
\item{\code{Clusters}}{A named factor representing cluster assignments for every
cell. Accessed with \code{\link{Clusters}}. Length should be equal to
number of cells in input data, and names should match colnames of gene
expression matrix (generally cell barcodes). Levels can be cluster numbers
or cluster names.}

\item{\code{ClustGeneStats}}{A named list of data frames, one entry for each level in
\code{Clusters} (with corresponding name). Accessed with
\code{ClustGeneStats}. Each entry is data frame containing gene summary
statistics for the cluster. Each data frame has the same number of rows as
the input gene expression matrix, where each row represents the results for
that gene (and shares its name).  The three variables in the data frame are
summary statistics. Detection Rate (\code{DR}) refers to the proprotion of
cells in the cluster in which gene cDNA was detected (a gene expression
value > 0). Mean Detected Gene Expression (\code{MDGE}) refers to the mean
normalized gene expression in cells from the cluster in which the gene was
detected (i.e. non-zero mean). Mean Gene Expression (\code{MGE}) is the
mean normalized gene expression of the gene in the cluster.}

\item{\code{DEvsRest}}{A named list of data frames, one entry for each level in
\code{Clusters} (with corresponding name). Accessed with \code{DEvsRest}.
Each entry is data frame containing gene differential expression stats when
comparing the cells of that cluster to all other cells in the input data,
as calculated by \code{CalcDEvsRest}. Rows represent genes, and variables
\strong{must include} \code{logGER} (an effect size measure: gene
expression ratio in log space, often referred to as logFC) and \code{FDR}
(significance measure: false discovery rate). This slot can be populated
using the function \code{\link{CalcDEvsRest}}, which can either calculate
differential expression for all clusters, or take an appropriately
formatted list of precomputed values. See \code{\link{CalcDEvsRest}}
documentation for details.}

\item{\code{DEcombn}}{A named list of data frames, one entry for each combination of
pairs of levels in \code{Clusters} (named as ClusterA-ClusterB). Accessed
with \code{DEcombn}. Each entry is a data frame containing gene
differential expression stats when comparing the two clusters, as
calculated by \code{\link{CalcDEvsCombn}}. Rows represent genes, and
variables \strong{must include} \code{logGER} (an effect size measure: gene
expression ratio in log space, often referred to as logFC) and \code{FDR}
(significance measure: false discovery rate). This slot can be populated
using the function \code{\link{CalcDEvsCombn}}, which can either calculate
differential expression for all combinations of clusters, or take an
appropriately formatted list of precomputed values. See
\code{\link{CalcDEvsCombn}} documentation for details.}

\item{\code{Silhouette}}{An object of class Silhouette defining a cluster
cohesion/separation statistic (silhouette width) for every cell. Accessed
with \code{Silhouette}. This slot can be populated using the function
\code{\link{CalcSilhouette}}, which is a wrapper to
\code{\link[cluster]{silhouette}} with distance calcuated using the same
reduced dimensional cell embedding as was used for clustering (see
\code{Param(sCVdata,"DRforClust")}).}

\item{\code{params}}{An object of class \code{\link{sCVparams}} containing the
parameters relevant to the data in this object. Accessed with \code{Param}.
Slots are accessed using \code{\link{Param}(sCVdata,slotName)}.}
}}

