% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sCVdataClass.R
\docType{class}
\name{sCVparams-class}
\alias{sCVparams-class}
\alias{sCVparams}
\title{Parameters used in the sCVdata class and methods.}
\description{
An S4 class storing parameters used in the sCVdata class and methods. Access
this object from the containing \code{\link{sCVdata}} object with
\code{\link{Param}}. Slots are accessed using
\code{\link{Param}(sCVdata,slotName)}
}
\section{Slots}{

\describe{
\item{\code{assayType}}{A length-one character vector representing the assay slot in
which the expression data is stored in the input object. This is not
required for Seurat v1 or v2 objects. See \code{\link{getExpr}} for
details.}

\item{\code{DRforClust}}{A length-one character vector representing the
dimensionality reduction method used as the input for clustering. This is
commonly PCA, and should correspond to the slot name of the cell embedding
in your input data - either the \code{type} argument in
\code{\link[SingleCellExperiment]{reducedDim}(x,type)} or the
\code{reduction.type} argument in
\code{\link[Seurat]{GetDimReduction}(object,reduction.type)} (v2) or
\code{reduction} in \code{\link[Seurat]{Embeddings}(object,reduction)}.}

\item{\code{exponent}}{A length-one numeric vector representing the base of the
log-normalized gene expression data to be processed. Generally gene
expression data is transformed into log2 space when normalizing (set this
to 2), though \code{Seurat} uses the natural log (set this to exp(1)).}

\item{\code{pseudocount}}{A length-one numeric vector representing the pseudocount
added to the data during log-normalization to avoid log(0) errors.
Generally this is 1.}

\item{\code{DRthresh}}{A length-one numeric vector between 0 and 1 representing the
detection rate threshold for inclusion of a gene in the differential
expression testing. A gene will be included if it is detected in at least
this proportion of cells in at least one of the clusters being compared.
Commonly set to 0.1.}
}}

\seealso{
\code{\link{sCVdata}} for containing class.
}
