% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadLabels2.R
\name{spreadLabels2}
\alias{spreadLabels2}
\title{Prevent overlapping labels}
\usage{
spreadLabels2(x, y, label, padding = 0, str.cex = par("cex"),
  str.font = par("font"))
}
\arguments{
\item{x}{The set of x coordinates for the points to plot}

\item{y}{The set of y coordinates for the points to plot}

\item{label}{A character vector of labels for each x,y pair.}

\item{padding}{A numeric value representing the extra proportion of string
size to add around each label. Not clear that this works as intended, so
leave it at zero.}

\item{str.cex}{The cex value for the label. Defaults to
\code{\link[graphics]{par}("cex")}. Passed to \code{\link[graphics]{strwidth}}.}

\item{str.font}{The font value for the label. Defaults to
\code{\link[graphics]{par}("font")}. Passed to \code{\link[graphics]{strwidth}}.}
}
\value{
A matrix with 2 columns corresponding to x and y coordinates of the
  label centre to be used downstream as an input e.g. to \code{text}
}
\description{
Determines clever placement of labels for plots made in base R
}
\references{
This code is inspired by
  \href{https://github.com/federicomarini/spreadLabels}{Federico Marini} and
  \href{https://github.com/ChristofferFlensburg/spreadLabels}{Christoffer
  Flensburg}.
}
