% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccess.R
\name{subsetCells}
\alias{subsetCells}
\title{Subset a data object to return just the specified cells}
\usage{
subsetCells(x, cells)
}
\arguments{
\item{x}{The single-cell data object.}

\item{cells}{A character vector of cell names to extract. This is 
intentionally less versatile than most subsetting, since older versions of 
Seurat required cell names explicitly.}
}
\description{
Internal fx.  Way shittier than the modern object class' subset fx. 
Req'd because old Seurat is a POS.
}
\details{
This is a wrapper function to the relevant class's subsetting method. 
Currently supported input object classes:
\itemize{
  \item Class \code{\link[Seurat]{seurat}/\link[Seurat]{Seurat}}.
  \item Class \code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}
}
