% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{tsne_labels}
\alias{tsne_labels}
\title{scClustViz plot element: Cluster names on cluster centroid.}
\usage{
tsne_labels(sCVd, cell_coord, lab_type)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{cell_coord}{A numeric matrix where named rows are cells, and two
columns are the x and y dimensions of the cell embedding.}

\item{lab_type}{One of "ClusterNames", "ClusterNamesAll", or "Clusters".
"ClusterNames" places cluster names (added to sCVdata object by
\code{\link{labelCellTypes}}) at the centroid of all points sharing that
cluster name (can span clusters). "ClusterNamesAll" places cluster names at
the centroid of each cluster. "Clusters" places cluster ID
(\code{levels(Clusters(sCVd))}) at the centroid of each cluster.}
}
\description{
See \code{\link{plot_tsne}} for application.
}
