% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{CalcCGS}
\alias{CalcCGS}
\alias{CalcCGS,sCVdata-method}
\title{Calculate cluster-wise gene statistics for sCVdata}
\usage{
CalcCGS(sCVd, inD)

\S4method{CalcCGS}{sCVdata}(sCVd, inD)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}
}
\value{
The function returns a list of dataframes. Each list element contains
  a named list of clusters at that resolution. Each of those list elements
  contains a dataframe where each sample is a gene, containing the following
  variables: \code{DR} is the proportion of cells in the cluster in which
  that gene was detected. \code{MDGE} is mean normalized gene expression for
  that gene in only the cells in which it was detected (see
  \code{\link{meanLogX}} for mean calculation). \code{MGE} is the mean
  normalized gene expression for that gene in all cells of the cluster (see
  \code{\link{meanLogX}} for mean calculation).
}
\description{
Calculates gene summary statistics per cluster for the clusters in an sCVdata
object, using the gene expression matrix from the input data object. This is
called by \code{\link{CalcSCV}} and you shouldn't need to call it on its own.
}
\details{
To help track its progress, this function uses progress bars from
\code{pbapply}. To disable these, set
\code{\link[pbapply]{pboptions}(type="none")}. To re-enable, set
\code{\link[pbapply]{pboptions}(type="timer")}.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Calculate cluster-wise gene stats for sCVdata
}}

\examples{
\dontrun{
ClustGeneStats(your_sCV_obj) <- CalcCGS(sCVd=your_sCV_obj,
                                        inD=your_scRNAseq_data_object)
}

}
\seealso{
\code{\link{CalcSCV}} for wrapper function to calculate all
  statistics for an sCVdata object, and \code{\link{fx_calcCGS}} for the
  internal function performing the calculations.
}
