% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{CalcDEvsRest}
\alias{CalcDEvsRest}
\alias{CalcDEvsRest,sCVdata-method}
\title{Calculates one vs. all DE tests for sCVdata}
\usage{
CalcDEvsRest(sCVd, inD, storeAllDE)

\S4method{CalcDEvsRest}{sCVdata}(sCVd, inD, storeAllDE)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}

\item{storeAllDE}{A logical vector of length 1 indicating whether to
calculate and store effect size information for all genes in the comparison
(TRUE), or just those passing the detection rate threshold for the Wilcoxon
rank-sum test (FALSE). Setting this to FALSE will reduce the size of the
output sCVdata object.}
}
\value{
A named list of data frames, one entry for each level in
  \code{Clusters(sCVd)} (with corresponding name).Each entry is data frame
  containing gene differential expression stats when comparing the cells of
  that cluster to all other cells in the input data. Rows represent genes,
  and variables include \code{logGER} (an effect size measure: gene
  expression ratio in log space, often referred to as logFC) and \code{FDR}
  (significance measure: false discovery rate). Also included are
  \code{Wstat} and \code{pVal}, the test statistic and the p-value of the
  Wilcoxon rank sum test.
}
\description{
Performs differential gene expression tests for each cluster in an sCVdata
object, comparing the cells in the cluster to the remaining cells in the data
using the gene expression matrix of input data object. Alternatively, this
function can be skipped, and existing DE test results can be assigned
directly to the sCVdata object.
}
\details{
This function performs Wilcoxon rank sum tests comparing gene expression
between each cluster and all other cells in the input data. Gene expression
ratio in log space (\code{logGER}) is reported for all genes in the
comparison. Genes are tested if they are detected in the cluster at a higher
proportion than \code{Param(sCVd,"DRthresh")}, and both unadjusted p-values
and false discovery rates are reported for all genes tested. To help track
its progress, this function uses progress bars from \code{pbapply}. To
disable these, set \code{\link[pbapply]{pboptions}(type="none")}. To
re-enable, set \code{\link[pbapply]{pboptions}(type="timer")}.

If using existing DE test results, assign results of one vs. all tests for
every cluster in sCVdata to the \code{\link{DEvsRest}} slot of the
\code{\link{sCVdata}} object. See example and slot documentation.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Calculate one vs. all DE tests for sCVdata
}}

\examples{
\dontrun{
## Example using CalcDEvsRest ##
DEvsRest(your_sCV_obj) <- CalcDEvsRest(sCVd=your_sCV_obj,
                                       inD=your_scRNAseq_data_object)


## Example using MAST results from Seurat to replace CalcDEvsRest ##
MAST_result <- Seurat::FindAllMarkers(your_seurat_obj,
                                      test.use="MAST",
                                      latent.vars="nUMI")

names(MAST_result)[names(MAST_result) == "avg_logFC"] <- "logGER"
# ^ Effect size variable must be named 'logGER'
names(MAST_result)[names(MAST_result) == "p_val_adj"] <- "FDR"
# ^ Significance variable must be named 'FDR'

MAST_result_list <- sapply(levels(MAST_result$cluster),
                           function(X)
                             MAST_result[MAST_result$cluster == X,],
                           simplify=F)
# ^ Must be named list with entries for every cluster.

DEvsRest(your_sCV_obj) <- MAST_result_list

}

}
\seealso{
\code{\link{CalcSCV}} for wrapper function to calculate all
  statistics for an sCVdata object,  and \code{\link{fx_calcESvsRest}} and
  \code{\link{fx_calcDEvsRest}} for the internal functions performing the
  calculations.
}
