% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{cosineSim}
\alias{cosineSim}
\title{Computes the cosine similarity between two vectors}
\usage{
cosineSim(A, B)
}
\arguments{
\item{A}{One vector in the pair to compare.}

\item{B}{The other vector in the pair to compare.}
}
\description{
Computes the cosine similarity between two vectors. Stolen from
\url{https://stats.stackexchange.com/q/31573}/
}
