% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{DEmarker}
\alias{DEmarker}
\alias{DEmarker,sCVdata-method}
\title{Find marker genes for each cluster}
\usage{
DEmarker(sCVd, FDRthresh)

\S4method{DEmarker}{sCVdata}(sCVd, FDRthresh)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of data frames, one entry for each cluster. Each data
  frame is the combined results of the set of pairwise gene expression tests
  comparing the cluster to each other cluster in the data. See
  \code{\link{CalcDEcombn}} and \code{\link{DEcombn}} for details.
}
\description{
Identifies the significantly positively differentially expressed genes
between a cluster and all other clusters, using the pairwise differential
gene expression testing results.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Find marker genes from sCVdata
}}

