% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\docType{methods}
\name{DEneighb}
\alias{DEneighb}
\alias{DEneighb,sCVdata-method}
\title{Find differentially expressed genes between nearest neighbouring clusters}
\usage{
DEneighb(sCVd, FDRthresh)

\S4method{DEneighb}{sCVdata}(sCVd, FDRthresh)
}
\arguments{
\item{sCVd}{An sCVdata object.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of data frames, one entry for each cluster. Each data
  frame is the results of the gene expression test comparing the cluster to
  its nearest neighbour. See \code{\link{CalcDEcombn}} and
  \code{\link{DEcombn}} for details.
}
\description{
Identifies the significantly differentially expressed genes between nearest
neighbouring clusters, using the pairwise differential gene expression
testing results.
}
\section{Methods (by class)}{
\itemize{
\item \code{sCVdata}: Find DE genes between nearest neighbours from sCVdata
}}

