% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcMarker}
\alias{fx_calcMarker}
\title{Internal function to find marker genes for each cluster}
\usage{
fx_calcMarker(deVS, FDRthresh)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of named numeric vectors, one entry for each cluster.
  Each vector is the FDR values for each gene that is significantly
  positively differentially expressed in that cluster over all other clusters
  in pairwise tests, with names corresponding to gene labels.
}
\description{
Internal function. See \code{\link{DEmarker}}.
}
\details{
Identifies the significantly positively differentially expressed genes
between a cluster and all other clusters, using the pairwise differential
gene expression testing results.
}
