% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deTest.R
\name{fx_calcNeighb}
\alias{fx_calcNeighb}
\title{Internal function to find DE genes between nearest neighbouring clusters}
\usage{
fx_calcNeighb(deVS, NN, FDRthresh)
}
\arguments{
\item{deVS}{List of pairwise DE results from \code{\link{CalcDEcombn}}.}

\item{NN}{A vector of nearest neighbours from \code{\link{DEdistNN}}.}

\item{FDRthresh}{False discovery rate for counting significance.}
}
\value{
A named list of named numeric vectors, one entry for each nearest
  neighbouring pair. Each vector is the FDR values for each gene that is
  significantly differentially expressed between that pair of clusters, with
  names corresponding to gene labels.
}
\description{
Internal function. See \code{\link{DEneighb}}.
}
\details{
Identifies the significantly differentially expressed genes between nearest
neighbouring clusters, using the pairwise differential gene expression
testing results.
}
