% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAccess.R
\name{getExpr}
\alias{getExpr}
\title{Get gene expression matrix from input data object}
\usage{
getExpr(x)
}
\arguments{
\item{x}{The single-cell data object.}
}
\description{
Extract the gene expression matrix from a single-cell data object containing
the input data for scClustViz visualization.
}
\details{
This is a wrapper function to the relevant class's normalized data slot
accessor method. Currently supported input object classes:
\itemize{
  \item Class \code{\link[Seurat]{seurat}} accessed by 
    \code{\link[Seurat]{GetAssayData}(x)}.
  \item Class \code{\link[SingleCellExperiment]{SingleCellExperiment}} 
    accessed by \code{\link[SingleCellExperiment]{logcounts}(x)}.
}
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}
}
