% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFx.R
\name{labelCellTypes}
\alias{labelCellTypes}
\title{scClustViz helper fx: Add predicted cell type names to cluster labels}
\usage{
labelCellTypes(sCV, cellMarkers, symbolMap)
}
\arguments{
\item{sCV}{An object of class \code{sCVdata}.}

\item{cellMarkers}{The \code{cellMarkers} argument from
\code{\link{runShiny}}. A list of marker genes for expected cell types.}

\item{symbolMap}{The output of \code{\link{map2symbol}}.}
}
\value{
Returns the sCVdata object with an added attribute
  '\code{ClusterNames}' to \code{Clusters(sCV)} containing the assigned cell
  type names for each cluster.
}
\description{
A bare-bones method of predicting cell types from marker genes.
}
\details{
Assigns cell type labels to each cluster based on the rank of median gene
expression of marker genes for each cell type. There are many more
intelligent methods for cell type prediction out there.  See
\href{https://doi.org/10.1101/521914}{CellAssign}, for example.
}
